🚚 Rename: 컴포넌트 파일들index.jsx로 변경
