+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from "react-router";
 import "./App.css";
-import Layout from "./components/Layout/Layout";
+import Layout from "./components/Layout";
 
 function App() {
   return (
@@ -13,7 +13,7 @@ function App() {
           <Route path="/post">
             <Route index element={/* 롤링페이퍼 생성페이지 */ null} />
             <Route path=":id/message" element={/* 롤링페이퍼 메세지 보내기 페이지 */ null} />
-            <Route element={<div>헤더서비스입니다</div>}>
+            <Route element={<null />}>
               <Route path=":id" element={/* 생성된 롤링페이퍼 페이지 */ null} />
               <Route path=":id/edit" element={/* 롤링페이퍼 수정 페이지 */ null} />
             </Route>
