+++ src/components/Layout/index.jsx
@@ -1,9 +1,10 @@
 import { Outlet } from "react-router";
+import Header from "../Header";
 
 const Layout = () => {
   return (
     <>
-      <header>헤더입니다!!!</header>
+      <Header />
       <Outlet />
     </>
   );
